move convert_to_sql_value to a different file to prevent circular imports
